Added a `Job.executor` property.
Since `Job.cancel()` requires the job to hold an executor instance,
this property is implied but not explicit. Since jobs are submitted
using `JobExecutor.submit()`, the job does not get to see an executor
instance unless the executor explicitly communicattes a reference to
itself as part of the `submit()` call. However, if the property is not
defined in the spec, an executor implementation cannot know what property
to set unless it looks at the specifics of the implementation of the `Job`
class. But that also prevents the executor implementation from being
portable to other implementations of the PSI/J core classes.